# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Forge 软件发布管理平台 API
  description: 本文档提供了 Forge 软件发布管理平台的详细 API 文档，包括认证、软件管理、统计分析等各个模块的接口说明。
  version: 1.0.0
  contact:
    name: Forge 开发团队
    email: dev@forge.com
servers:
  - url: http://localhost:5000/api
    description: 开发服务器
  - url: https://api.forge.com/api
    description: 生产服务器
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - 认证
      summary: 管理员登录
      description: 用户登录获取访问令牌
      security: []  # 无需认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: 用户名
                  example: admin
                password:
                  type: string
                  description: 密码
                  example: admin123
                  format: password
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 登录成功
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        description: 访问令牌
                        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
                      refresh_token:
                        type: string
                        description: 刷新令牌
                        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/refresh:
    post:
      tags:
        - 认证
      summary: 刷新令牌
      description: 使用刷新令牌获取新的访问令牌
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 令牌刷新成功
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        description: 新的访问令牌
                        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/profile:
    get:
      tags:
        - 认证
      summary: 获取用户信息
      description: 获取当前登录用户的信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取用户信息成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取用户信息成功
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/change-password:
    post:
      tags:
        - 认证
      summary: 修改密码
      description: 修改当前登录用户的密码
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                  description: 原密码
                  example: old_password
                  format: password
                new_password:
                  type: string
                  description: 新密码
                  example: new_password
                  format: password
      responses:
        '200':
          description: 密码修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/init-admin:
    post:
      tags:
        - 认证
      summary: 初始化管理员账户
      description: 首次使用时初始化管理员账户（仅当没有管理员时可用）
      security: []  # 无需认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: 用户名
                  example: admin
                password:
                  type: string
                  description: 密码
                  example: admin123
                  format: password
                email:
                  type: string
                  description: 邮箱
                  example: admin@example.com
                  format: email
      responses:
        '200':
          description: 管理员创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 管理员创建成功
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'

  # 软件管理相关接口
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token认证
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
          example: 1
        username:
          type: string
          description: 用户名
          example: admin
        email:
          type: string
          description: 邮箱
          example: admin@example.com
          format: email
        role:
          type: string
          description: 用户角色
          example: admin
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: 2023-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: 2023-01-01T00:00:00Z
...
  responses:
    Success:
      description: 操作成功
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: 操作成功
              data:
                type: object
                nullable: true
                example: null
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 请求参数错误
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: INVALID_PARAMETER
                  message:
                    type: string
                    example: 请求参数无效
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 未授权，请登录
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: UNAUTHORIZED
                  message:
                    type: string
                    example: 未授权访问
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 禁止访问
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: 权限不足
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 资源不存在
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: RESOURCE_NOT_FOUND
                  message:
                    type: string
                    example: 资源不存在
...